added clearer examples
