--- lib/wraith/javascript/phantom--nojs.js
@@ -1,6 +1,0 @@
-var system = require('system');
-
-require('./_phantom__common.js')({
-  systemArgs:        system.args,
-  javascriptEnabled: false
-});
