--- templates/configs/component.yaml
@@ -1,60 +1,0 @@
-# Add as many domains as necessary. Key will act as a label
-domains:
-  english: "http://www.bbc.co.uk"
-
-# (optional) JavaScript file to execute before taking screenshot of every path.
-# Can be used to reload current page with headers specified, or to dynamically
-# AJAX in content, or all manner of other uses!
-#before_capture: 'javascript/global.js'
-
-# Type page URL paths below, here are a couple of examples
-paths:
-  clickable_guide:
-    path: /news/entertainment-arts-27221191
-    selector: '.idt__news' # selector to take a screenshot of
-  clickable_guide__after_click:
-    path: /news/entertainment-arts-27221191
-    selector: '.idt__news'
-    # (optional) JavaScript file to execute before taking the screenshot of this path.
-    # This is always executed AFTER the 'global' before_capture hook toward the top of this file.
-    before_capture: 'javascript/beforeCapture--casper_example.js'
-
-# amount of fuzz ImageMagick will use when comparing images. A higher fuzz makes the comparison less strict.
-fuzz: '20%'
-
-# the maximum acceptable level of difference (in %) between two images.
-# Wraith considers it a failure if an image diff goes above this threshold.
-threshold: 5
-
-# screen widths (and optional height) to resize the browser to before taking the screenshot
-screen_widths:
-  - 320
-  - 600x768
-  - 768
-  - 1024
-  - 1280
-
-# resize to each screen width, or reload at each screen width. (The former is more efficient).
-resize_or_reload: 'resize'
-
-# the engine to run Wraith with.
-browser: "casperjs"
-
-# the directory that your base screenshots will be stored in
-history_dir: 'shots_base'
-
-# the directory that your latest screenshots will be stored in
-directory: 'shots'
-
-gallery:
-  template: 'slideshow_template' # 'basic_template' (default)
-  thumb_width: 200
-  thumb_height: 200
-
-# choose how results are displayed in the gallery (default is `alphanumeric` if omitted)
-# Different screen widths are always grouped together.
-# Options:
-#   alphanumeric - all paths (with or without a difference) are shown, sorted by path
-#   diffs_first - all paths (with or without a difference) are shown, sorted by difference size (largest first)
-#   diffs_only - only paths with a difference are shown, sorted by difference size (largest first)
-mode: alphanumeric
